+++ bitcodin/api.py
@@ -350,3 +350,46 @@ def create_thumbnail(thumbnail_request):
 
     thumbnail_response = BitcodinObject(res)
     return thumbnail_response
+
+
+def create_transmux_job(transmux_request):
+    """
+    Transmux a Job to several output files for progressive streaming
+    :param transmux_request:
+    :return: string
+    """
+
+    url = get_api_base() + '/transmuxjob'
+    res = RestClient.post(url=url, headers=create_headers(),
+                          content=transmux_request.to_json())
+
+    transmux_job_response = BitcodinObject(res)
+    return transmux_job_response
+
+
+def get_transmux_job(transmux_job_id):
+    """
+    Gets details about a transmuxing job
+    :param transmux_job_id: string
+    :return: string
+    """
+
+    url = get_api_base() + '/transmuxjob/%d' % transmux_job_id
+    res = RestClient.get(url=url, headers=create_headers())
+
+    transmux_job_response = BitcodinObject(res)
+    return transmux_job_response
+
+
+def get_manifest_info(job_id):
+    """
+    Gets manifest info of a job
+    :param job_id:
+    :return: string
+    """
+
+    url = get_api_base() + '/job/%d' % job_id + '/manifest-info'
+    res = RestClient.get(url=url, headers=create_headers())
+
+    manifest_info_response = BitcodinObject(res)
+    return manifest_info_response